+++ Security/Authentication/Provider/WsseProvider.php
@@ -14,7 +14,7 @@ use Symfony\Component\Security\Core\User\UserProviderInterface;
 
 /**
  * 
- * @author cedrick
+ * @author Cedrick Oka Baidai <okacedrick@gmail.com>
  * 
  */
 class WsseProvider implements AuthenticationProviderInterface