+++ Security/Guard/JwtGuardAuthenticator.php
@@ -16,8 +16,9 @@ use Symfony\Component\Security\Guard\AbstractGuardAuthenticator;
 
 /**
  * 
- * @author cedrick
+ * @author Cedrick Oka Baidai <okacedrick@gmail.com>
  * 
+ * @deprecated 1.2.0 Not using
  */
 class JwtGuardAuthenticator extends AbstractGuardAuthenticator
 {